feature: add status classes
