+++ src/Campaigns/resources/admin/components/CampaignDetailsPage/CampaignDetailsPage.module.scss
@@ -82,16 +82,27 @@
     & > span {
         padding-inline: 0.5rem;
         padding-block: 0.185rem;
-        background: #f49420;
-        border-radius: 5px;
+        border-radius: 1rem;
         font-size: 0.7rem;
         font-weight: 700;
-        text-transform: uppercase;
+        text-transform: capitalize;
         letter-spacing: 0.04rem;
         transform: translateY(0.075rem);
         color: #ffffff;
     }
+
+    .draftStatus {
+        background: #FFDAA6;
+        color: #1A0F00;
+
+    }
+
+    .activeStatus {
+        background: #CEF2CF;
+        color: #022603;
+    }
 }
+
 
 .flexRow:not(:first-child) {
     flex: 1;
